Added Exercise 2
You will know find the second exercise in "chapter2" - it contains the R
code and the results of the executed code as well as my interpretations
of the results.
Also, there are the data files in .csv and .txt format and the "Week 2,
Data Analysis.R file with the code for the data analysis.